backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
google-genai>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
Pillow>=10.2.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
